LayoutTests/imported/w3c/web-platform-tests/web-animations/interfaces/Animation/commitStyles-expected.txt


PASS Throws if the target element is not something with a style attribute
PASS Throws if the target element is a pseudo element
PASS Checks the pseudo element condition before the not rendered condition
PASS Throws if the target effect is disconnected
PASS Throws if the target effect is display:none
PASS Throws if the target effect's ancestor is display:none
PASS Treats display:contents as rendered
PASS Treats display:contents in a display:none subtree as not rendered
FAIL Commits styles assert_approx_equals: expected 0.2 +/- 0.0001 but got 0.1
FAIL Commits styles for backwards animation assert_approx_equals: expected 0.5 +/- 0.0001 but got 0.1
PASS Commits values calculated mid-interval
PASS Commits values during the start delay
PASS Commits value after the active interval
FAIL Commits the intermediate value of an animation in the middle of stack assert_approx_equals: expected 0.4 +/- 0.0001 but got 0.2
FAIL Commits the intermediate value of an animation up to the middle of the stack assert_approx_equals: expected 0.4 +/- 0.0001 but got 0.1
FAIL Commits styles for an animation that has been removed assert_approx_equals: expected 0.2 +/- 0.0001 but got 0.1
FAIL Commit composites on top of the underlying value assert_approx_equals: expected 0.5 +/- 0.0001 but got 0.3
FAIL Commits logical properties assert_equals: expected "20px" but got "10px"
FAIL Commits logical properties as physical properties assert_equals: expected "20px" but got "10px"
FAIL Commits em units as pixel values assert_approx_equals: expected 100 +/- 0.0001 but got 784
FAIL Commits custom variables assert_approx_equals: expected 0.8 +/- 0.0001 but got 0.5
FAIL Commits variable references as their computed values assert_approx_equals: expected 0.5 +/- 0.0001 but got 1
FAIL Commits relative line-height assert_approx_equals: expected 15 +/- 0.0001 but got NaN
FAIL Commits transforms assert_equals: expected "matrix(1, 0, 0, 1, 20, 20)" but got "none"
FAIL Commits styles for individual transform properties assert_equals: expected "200px" but got "100px"
FAIL Commits transforms as a transform list assert_equals: expected "translate(20px, 20px)" but got "none"
FAIL Commits matrix-interpolated relative transforms assert_equals: Resolved transform is correct after commit. expected "matrix(2, 0, 0, 2, 100, 0)" but got "matrix(2, 0, 0, 2, 0, 0)"
PASS Commits 'none' transform
FAIL Commits shorthand styles assert_equals: expected "20px" but got "10px"
FAIL Triggers mutation observers when updating style assert_equals: Should have one mutation record expected 1 but got 0
PASS Does NOT trigger mutation observers when the change to style is redundant
